Make user mapping more flexible
This allows you to specify what uids to use from the host, using the
key `RemapUidRanges`, either by giving a set of ranges, or by giving
a username which looks up the subuids for that user.

Additionally you can specify the first remapped uid with
`RemapUidStart`, which defaults to 1. This means we have the same
behaviour as before, but allows it to change.

Also change NoUsermap to RemapUsers (with opposite default value), to
avoid negative keynames.