unitfile: Add quad_unit_file_lookup_all_strv()
This function collects all the values for a given key and splits them
by whitespace similar to the systemd function config_parse_strv,
merging the results.

This is useful for parsing any kind of multi-arg option with similar
semantics. In particular it is what the Alias, WantedBy and RequiredBy
keys in the [Install] group needs.